requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.3
cython<3.0.0
colorama==0.4.6
flask-jwt-extended==4.5.2
flask-restful==0.3.10
flask-sqlalchemy==3.0.5
flask==2.3.2
greenlet==2.0.2
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
pyjwt==2.7.0
pytz==2023.3
pyyaml~=6.0
six==1.16.0
sqlalchemy==2.0.17
typing-extensions==4.6.3
waitress==2.1.2
werkzeug==2.3.6
